Difficult to estimate the benefit: when we're faster we have less messages so less removals.
Here are a few tests: simul 180 nodes, threshold 178, 2 dead nodes
**before**
8.2s   7400 removed
7.5s 8600 removed
9.5s 9600 removed
8.5s 7800 removed
11.5s 12000 removed

**after**
7.8s  11000 removed
6s 6300 removed
8.5s 6500 removed
7.2s 9600 removed
5.7s 5800 removed

So it seems faster with the new version