@nkeywal 2019-01-16
Yep it seems better even if we will have to spend more time on this. Let's commit for now.
Thanks for the review!